% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_views.R
\name{bottomleft}
\alias{bottomleft}
\title{Print bottom-right of matrix}
\usage{
bottomleft(M, nView = 5)
}
\arguments{
\item{M}{matrix to view}

\item{nView}{Number of rows/cols to view}
}
\value{

}
\description{
Print bottom-right of matrix
}
\examples{
M <- diag(1:12)
bottomright(M)
bottomright(M, 3)
}
