% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_views.R
\name{bottomright}
\alias{bottomright}
\title{Print bottomleft of matrix}
\usage{
bottomright(M, nView = 5)
}
\arguments{
\item{M}{matrix to view}

\item{nView}{Number of rows/cols to view}
}
\description{
Print bottomleft of matrix
}
\examples{
M <- diag(1:12)
bottomleft(M)
bottomleft(M, 3)
}
