% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causality_check.R
\name{causality_check}
\alias{causality_check}
\title{Check causality of VAR}
\usage{
causality_check(phi)
}
\arguments{
\item{phi}{"rbind" of Phi_1,...,Phi_p matricies}
}
\value{
TRUE if all roots of companion matrix are less than 1
}
\description{
Check causality of VAR
}
