% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_ts.R
\name{compare_ts}
\alias{compare_ts}
\title{Compare 2 time series
Plot standardized time series together along with their difference}
\usage{
compare_ts(x1, x2)
}
\arguments{
\item{x1}{first time series}

\item{x2}{second time series}
}
\value{
a plot
}
\description{
Compare 2 time series
Plot standardized time series together along with their difference
}
