% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_time_events.R
\name{count_time_events}
\alias{count_time_events}
\title{Count time-stamped events}
\usage{
count_time_events(event_dates, event_type = c("year", "month", "week",
  "day", "hour", "minute"), event_all_possible = NULL)
}
\arguments{
\item{event_dates}{vector of class "Date" of events to be counts}

\item{event_type}{'year', 'month', 'week', 'day', 'hour' or 'minute'}

\item{event_all_possible}{(optional) specify all possible dates to count}
}
\value{
table of counts for each event by event_type
}
\description{
Count time-stamped events
}
