% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand-grids.R
\name{expand.grids}
\alias{expand.grids}
\title{expand.grid() function for increasing dimension d
mimics $d$ nested for loops with elements from $x$.}
\usage{
expand.grids(x, d)
}
\arguments{
\item{x}{vector of elements to be looped over}

\item{d}{number of nested for loops}
}
\value{
d x length(x) matrix with each row giving the next iteration of loop
}
\description{
expand.grid() function for increasing dimension d
mimics $d$ nested for loops with elements from $x$.
}
