% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastMerge.R
\name{fastmerge}
\alias{fastmerge}
\title{Data.frame merge
Merge two dataframes based on column names (not column order)}
\usage{
fastmerge(d1, d2)
}
\arguments{
\item{d1}{first data.frame}

\item{d2}{second data.frame}
}
\value{
merged data.frame
}
\description{
Data.frame merge
Merge two dataframes based on column names (not column order)
}
