% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tri-matrix.R
\name{fill_upper_tri}
\alias{fill_upper_tri}
\title{fill the upper triangle of a (d x d) matrix with the vector v}
\usage{
fill_upper_tri(v, d)
}
\arguments{
\item{v}{vector with choose(d,2) elements to fill upper triangle}

\item{d}{dimensionality of (dxd) matrix output}
}
\value{
a (dxd) upper triangluar matrix filled with v
}
\description{
fill the upper triangle of a (d x d) matrix with the vector v
}
