% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sigma.R
\name{gen_Sigma}
\alias{gen_Sigma}
\title{Random covariance matrix generation. This routine uses either the method
of diagonal dominance or an autoregressive time series structure. The
Diagnoal dominance does not produce large correlations but a wider variety
or struture.}
\usage{
gen_Sigma(d, ar1 = FALSE, phi = 0.5)
}
\arguments{
\item{d}{The dimensionality of the desired Covariance matrix}

\item{ar1}{The method that should be used. If FALSE(default) then
diagnal dominance is used. If TRUE then an AR time series matrix used.}

\item{phi}{The AR parameter. Only needed or used when ar1=TRUE}
}
\value{
Sigma A positive definate Covariance matrix.
}
\description{
Random covariance matrix generation. This routine uses either the method
of diagonal dominance or an autoregressive time series structure. The
Diagnoal dominance does not produce large correlations but a wider variety
or struture.
}
\examples{
gen_sigma(5)


}
