% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not-in.R
\name{\%!in\%}
\alias{\%!in\%}
\title{Not in}
\usage{
x \%!in\% y
}
\arguments{
\item{x}{value to be checked}

\item{y}{values to check over}
}
\value{
TRUE/FALSE opposite of %in%
}
\description{
The converse of %in%
}
