% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_print.R
\name{hist_print}
\alias{hist_print}
\title{Nice printing of histogram bins along with counts}
\usage{
hist_print(x, breaks = "Sturges")
}
\arguments{
\item{x}{vector of values}

\item{breaks}{optional arguement to pass to hist()}
}
\value{
data.frame with two columns, ranges and counts.
}
\description{
Nice printing of histogram bins along with counts
}
\examples{
x <- rnorm(100)
hist_print(x)
hist_print(x, breaks = c(-Inf, -4, -2, 0, 2, 4, Inf))
}
