% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_package_functions.R
\name{list_functions}
\alias{list_functions}
\title{List all functions exported when loading a package}
\usage{
list_functions(package.name)
}
\arguments{
\item{package.name}{characer string giving the name of the package}
}
\value{
character vector with names of functions exported
}
\description{
List all functions exported when loading a package
}
