% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrbind.R
\name{lrbind}
\alias{lrbind}
\title{list row bind}
\usage{
lrbind(List)
}
\arguments{
\item{List}{list where each object is a data.frame}
}
\value{
single data.frame where each object is row bound
}
\description{
list row bind
}
