% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_head.R
\name{matrix_head}
\alias{matrix_head}
\title{View upper left head of a matrix.
Analog to base R head() function for matrix}
\usage{
matrix_head(M, n = 10)
}
\arguments{
\item{M}{matrix}

\item{n}{number of rows/cols to print (default is 10)}
}
\value{
prints M[1:n, 1:n]
}
\description{
View upper left head of a matrix.
Analog to base R head() function for matrix
}
