% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_directory.R
\name{parse_for_error}
\alias{parse_for_error}
\title{Parse directory of R scripts for errors
Check which file in a directory throws and error}
\usage{
parse_for_error(path.to.directory)
}
\arguments{
\item{path.to.directory}{character string giving path to directory to be
checked}
}
\value{
output
}
\description{
Parse directory of R scripts for errors
Check which file in a directory throws and error
}
