% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_airline.R
\name{sim_airline}
\alias{sim_airline}
\title{Simulate from SARIMA(0 1 1)(0 1 1) model with integer periodicity}
\usage{
sim_airline(
  n = 120,
  s = 12,
  theta = 0.75,
  seasonal_theta = 0.75,
  sigma = 1,
  burn = 10^4,
  demean = TRUE,
  first_obs_zero = TRUE,
  start_date = c(2000, 1)
)
}
\arguments{
\item{n}{length of simulated series}

\item{s}{frequency of series (eg 12 for monthly or 4 for quarterly)}

\item{theta}{non-seasonal moving average parameter}

\item{seasonal_theta}{seasonal moving average parameter}

\item{sigma}{standard deviation of the innovations}

\item{burn}{length of burn in to begin series}

\item{demean}{boolean, should the series be demeaned? (default TRUE)}

\item{first_obs_zero}{boolean, should the series be scaled so the first
observation is at zero? Overides demean value.}

\item{start_date}{input to ts() argument start}
}
\value{
simulated series as a ts object
}
\description{
Simulate from SARIMA(0 1 1)(0 1 1) model with integer periodicity
}
\examples{
y = sim_airline(theta = .9, seasonal_theta = .9)
}
