% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourceAll_from_subdirectory.R
\name{sourceAll_from_subdirectory}
\alias{sourceAll_from_subdirectory}
\title{Source all files in a subdirector of current working directory
Useful for IRE submission.}
\usage{
sourceAll_from_subdirectory(subDir_with_slash)
}
\arguments{
\item{subDir_with_slash}{subdirectory name}
}
\description{
Source all files in a subdirector of current working directory
Useful for IRE submission.
}
