% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_views.R
\name{topleft}
\alias{topleft}
\title{Print topleft of matrix}
\usage{
topleft(M, nView = 5)
}
\arguments{
\item{M}{matrix to view}

\item{nView}{Number of rows/cols to view}
}
\value{
print matrix
}
\description{
Print topleft of matrix
}
\examples{
M <- diag(1:12)
topleft(M)
topleft(M, 3)
}
