% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_views.R
\name{topright}
\alias{topright}
\title{Print topright of matrix}
\usage{
topright(M, nView = 5)
}
\arguments{
\item{M}{matrix to view}

\item{nView}{Number of rows/cols to view}
}
\value{

}
\description{
Print topright of matrix
}
\examples{
M <- diag(1:12)
topright(M)
topright(M, 3)
}
