% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tri-matrix.R
\name{upTri}
\alias{upTri}
\title{upper triangluar elements of matrix}
\usage{
upTri(A)
}
\arguments{
\item{A}{square matrix}
}
\value{
vector of upper triangluar elements of input matrix
}
\description{
upper triangluar elements of matrix
}
