% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withErrorTracing.R
\name{withErrorTracing}
\alias{withErrorTracing}
\title{Not sure what this function does}
\usage{
withErrorTracing(expr, silentSuccess = FALSE)
}
\arguments{
\item{expr}{an expression?}

\item{silentSuccess}{???}
}
\value{
not sure
}
\description{
Not sure what this function does
}
