% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize}
\alias{standardize}
\title{Subtract mean and divide by standard deviation}
\usage{
standardize(x)
}
\arguments{
\item{x}{vector to be standardized}
}
\value{
vector of same length of x with mean subtracted and divided 
            through by the variance
}
\description{
Subtract mean and divide by standard deviation
}
