% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tri-matrix.R
\name{upTri}
\alias{upTri}
\title{Upper triangluar elements of matrix}
\usage{
upTri(A, byrow = FALSE)
}
\arguments{
\item{A}{square matrix}

\item{byrow}{boolean (default=FALSE) if returned vector should be by row.}
}
\value{
ector of upper triangluar elements of input matrix
}
\description{
Upper triangluar elements of matrix
}
